Support negative n-best, which samples from all hypotheses
Fixes #73.

+++ sentencepiece-sys/src/ffi/sentencepiece.cpp
@@ -75,9 +75,9 @@ unsigned char *spp_encode_as_serialized_proto(SentencePieceProcessor *spp, char 
   return data;
 }
 
-unsigned char *spp_sample_encode_as_serialized_proto(SentencePieceProcessor *spp, char const *sentence, size_t sentence_len, size_t *len, size_t nbest, float alpha) {
+unsigned char *spp_sample_encode_as_serialized_proto(SentencePieceProcessor *spp, char const *sentence, size_t sentence_len, size_t *len, int nbest, float alpha) {
   auto sentence_view = absl::string_view(sentence, sentence_len);
-  auto serialized = spp->SampleEncodeAsSerializedProto(sentence_view, static_cast<int>(nbest), alpha);
+  auto serialized = spp->SampleEncodeAsSerializedProto(sentence_view, nbest, alpha);
 
   *len = serialized.size();
   unsigned char *data = (unsigned char *) malloc(serialized.size());
